+++ Changelog.md
@@ -354,7 +354,7 @@ Grammar/parser now compiles, but the insertion point logic is broken: My declara
 - [grammar, parser] Replace assignmentRvalue with expression in assignmentStatement.  Use declareAssignmentExpression as expression alias, but just in the declaration rules (to distinguish assignment from initialization.)
   * more tests: `init_expr_unary initlisttrunc initsequence init_expr_bool forward_ref_init init_expr_call`
   * add placeholder test: `error_nonconst_init.silly` -- doesn't fail, but I want it to (see TODO)
-
+  * `perl -p -i -e 's/MLIRListener/ParseListener/g'` -- Bad name: It's an MLIR builder or Silly Parse Listener, but not an MLIRListener
 
 ## tag: V7 (Jan 4, 2025)